+++ rust/cloud-storage/macro_event_broker/src/domain/service/test.rs
@@ -54,7 +54,7 @@ enum ExampleTopicEvent {
 impl TopicEvent for ExampleTopicEvent {
     type Topic = MacroExampleTopic;
 
-    fn schema_version(&self) -> u16 {
+    fn schema_version(&self) -> u8 {
         1
     }
 }
